refactor: UI 및 메인 화면 출력 정상화
