--- src/Order/components/BottomPurchaseBar/BottomPurchaseBar.styles.ts
@@ -1,14 +1,0 @@
-import styled from '@emotion/styled';
-import { theme } from '@/styles/theme';
-
-export const Footer = styled.div`
-  position: sticky;
-  bottom: 0;
-  z-index: 100;
-  align-items: center;
-  background-color: ${theme.colors.semanticColor.brandColor.kakaoYellow};
-`;
-
-export const PriceText = styled.p`
-  ${theme.typography.title2Bold};
-`;
