+++ src/features/Gift/components/TrendRanking/TrendRankingStyle.ts
